Some manual validations added.

+++ Para.Api/Para.Bussiness/Command/Customer/DeleteCustomer/DeleteCustomerCommandHandler.cs
@@ -2,6 +2,7 @@
 using MediatR;
 using Para.Base.Response;
 using Para.Data.UnitOfWork;
+using Para.Schema;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -23,6 +24,10 @@ namespace Para.Bussiness.Command.Customer.DeleteCustomer
 
         public async Task<ApiResponse> Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
         {
+            if (request.CustomerId <= 0)
+            {
+                return new ApiResponse("Invalid Customer Id");
+            }
             await unitOfWork.CustomerRepository.Delete(request.CustomerId);
             await unitOfWork.Complete();
             return new ApiResponse();
